# docs/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraceMind Controller Server API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/workspaces/init": {
      "post": {
        "tags": [
          "workspace"
        ],
        "summary": "Init Workspace Route",
        "operationId": "init_workspace_route_api_v1_workspaces_init_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/mount": {
      "post": {
        "tags": [
          "workspace"
        ],
        "summary": "Mount Workspace",
        "operationId": "mount_workspace_api_v1_workspaces_mount_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MountWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/select": {
      "post": {
        "tags": [
          "workspace"
        ],
        "summary": "Select Workspace",
        "operationId": "select_workspace_api_v1_workspaces_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/llm/configs": {
      "get": {
        "tags": [
          "llm"
        ],
        "summary": "List Configs",
        "operationId": "list_configs_api_v1_llm_configs_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace that owns the config",
              "title": "Workspace Id"
            },
            "description": "Workspace that owns the config"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LlmConfigInfo"
                  },
                  "title": "Response List Configs Api V1 Llm Configs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Create Config",
        "operationId": "create_config_api_v1_llm_configs_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace that owns the config",
              "title": "Workspace Id"
            },
            "description": "Workspace that owns the config"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmConfigCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmConfigInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LlmConfigCreateRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "prompt_template_version": {
            "type": "string",
            "title": "Prompt Template Version"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "model",
          "prompt_template_version"
        ],
        "title": "LlmConfigCreateRequest"
      },
      "LlmConfigInfo": {
        "properties": {
          "config_id": {
            "type": "string",
            "title": "Config Id"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "prompt_template_version": {
            "type": "string",
            "title": "Prompt Template Version"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "config_id",
          "model",
          "prompt_template_version",
          "prompt_version",
          "created_at"
        ],
        "title": "LlmConfigInfo"
      },
...
      "MountWorkspaceRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "MountWorkspaceRequest"
      },
...
      "SelectWorkspaceRequest": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          }
        },
        "type": "object",
        "required": [
          "workspace_id"
        ],
        "title": "SelectWorkspaceRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceInfo": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "root": {
            "type": "string",
            "title": "Root"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "directories": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Directories"
          },
          "commit_policy": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Commit Policy"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "name",
          "root",
          "languages",
          "directories",
          "commit_policy"
        ],
        "title": "WorkspaceInfo"
      },
      "WorkspaceInitRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "languages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Languages"
          },
          "directories": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directories"
          },
          "commit_policy": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Policy"
          },
          "append_gitignore": {
            "type": "boolean",
            "title": "Append Gitignore",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "path",
          "name"
        ],
        "title": "WorkspaceInitRequest"
      },
      "WorkspaceInitResponse": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "manifest_path": {
            "type": "string",
            "title": "Manifest Path"
          },
          "directories": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Directories"
          },
          "sample_controller_bundle": {
            "type": "string",
            "title": "Sample Controller Bundle"
          },
          "sample_intent": {
            "type": "string",
            "title": "Sample Intent"
          },
          "gitignore_snippet": {
            "type": "string",
            "title": "Gitignore Snippet"
          },
          "gitignore_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gitignore Path"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "name",
          "manifest_path",
          "directories",
          "sample_controller_bundle",
          "sample_intent",
          "gitignore_snippet",
          "gitignore_path"
        ],
        "title": "WorkspaceInitResponse"
      }
    }
  }
}